+++ app/scripts/configs/tracks-info.js
@@ -1622,7 +1622,9 @@ export const TRACKS_INFO = [
       'fontSize',
       'fontIsLeftAligned',
       'showMousePosition',
-      'mousePositionColor'
+      'mousePositionColor',
+      'tickPositions',
+      'tickFormat'
     ],
     defaultOptions: {
       color: '#808080',