+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Next
 
 - Prettified JS code
-- Added `labelShowAssembly` as an option to allow hiding the `hg19 | ` text
+- Added `labelShowAssembly` as an option to allow hiding the assembly in track options `hg19 | ` text
+- Added `tickFormat` and `tickPosition` options to the chromosome labels track
 - Enabled the colorbar slider by adding the options `colorbarPosition` and `colorbarBackgroundColor` for the `horizontal-multivec`
 - Added release notes to docs.